+++ packages/expo-ui/android/src/main/java/expo/modules/ui/ExpoUIModule.kt
@@ -34,5 +34,7 @@ class ExpoUIModule : Module() {
         "onExpandedChanged"
       )
     }
+
+    View(ProgressView::class)
   }
 }